kinda
he philanthropy project of Clinton finantial group was used for the homicide of Akishinomiya, but where is the political funds from?they can not refuse it if asked from America. Even if murder, it is the LDP. Past party.Japan
and also Clinton financial group also made a request for homicide to the Japanese government. the group that should be crushed. It was more darkness deal that it was murder.i forgot. so it's strange. Mrs Clinton did not say anything to a conglomerate, so the Clinton financial group of oneself said that there was no reason for killing me.
If this keeps going as it is, Article 9 amendment. Japanese politicians are making political contributions to Clinton financial group. Japan is shielded by Russia. and Japan force attack to north Korea under Princess Diana in heaven. Japan is a facto military power country nationalism. For this reason Japan discards democracy. And we become border guards. The Abe 's Korean drama is already annoying to us which the system is already finished. so we would be become an orphan of the world.
what a this story is! this is jet-black sunrise.An international monitoring violation occurred in the United States